Extract interface & handle cases when video_id is already present in DB
